Update server.c
this is what I was thinking for the server side, since we have to create the socket, bind it try to open the imagefile.img, I was thinking could point to the char array, hostname[] that you defined on the client side, and then we can read into the buffer, and send data to the client. I read in the homework that upon successful connection to the client a handler similar to an fd, is sent and needs to be used as a reference, so I put in some of the methods  I think might work, if I am thinking correctly as how this is supposed to go.